# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Recipe:
    properties:
      id:
        description: '`swaggerignore:"true"`'
        type: string
      ingredients:
        items:
          type: string
        type: array
      instructions:
        items:
          type: string
        type: array
      name:
        type: string
      publishedAt:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /auth:
    get:
      consumes:
      - application/json
      description: Login with username and password
      parameters:
      - description: password
        in: query
        name: password
        required: true
        type: string
      - description: username
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            type: string
        "401":
          description: Invalid input
          schema:
            type: string
      summary: /siginin
      tags:
      - auth
    post:
      consumes:
      - application/json
      description: Get new token in exchange for an old one
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            type: string
        "400":
          description: Token is new and doesn't need a refresh
          schema:
            type: string
        "401":
          description: Invalid credentials
          schema:
            type: string
      summary: /refresh
      tags:
      - auth
  /recipes:
    get:
      consumes:
      - application/json
      description: Returns list of recipes
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            items:
              $ref: '#/definitions/models.Recipe'
            type: array
      summary: recipes listRecipes
      tags:
      - recipe
    post:
      consumes:
      - application/json
      description: create an rew recipe
      parameters:
      - description: Recipe Schema
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/models.Recipe'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Invalid input
          schema:
            type: string
      summary: recipes newRecipe
      tags:
      - recipe
  /recipes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing recipe
      parameters:
      - description: ID of the recipe
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Recipe'
        "404":
          description: Invalid recipe ID
          schema:
            type: string
      summary: recipes deleteRecipe
      tags:
      - recipe
    get:
      consumes:
      - application/json
      description: Get an one existing recipe
      parameters:
      - description: ID of the recipe
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Recipe'
        "404":
          description: Invalid recipe ID
          schema:
            type: string
      summary: recipes get one Recipe
      tags:
      - recipe
    put:
      consumes:
      - application/json
      description: Update an existing recipe
      parameters:
      - description: ID of the recipe
        in: path
        name: id
        required: true
        type: string
      - description: Recipe Schema
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/models.Recipe'
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Invalid recipe ID
          schema:
            type: string
      summary: recipes updateRecipe
      tags:
      - recipe
  /recipes/recipe/:id:
    get:
      consumes:
      - application/json
      description: Search recipes based on tags
      parameters:
      - description: ID of recipe
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Recipe'
        "404":
          description: Invalid recipe ID
          schema:
            type: string
      summary: recipes findRecipe
      tags:
      - recipe
  /recipes/search/:tag:
    get:
      consumes:
      - application/json
      description: Search recipes based on tags
      parameters:
      - description: recipe tag
        in: query
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/models.Recipe'
      summary: recipes findRecipe
      tags:
      - recipe
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"